リポジトリ検索をコンポーネントとして切り出し
