+++ renderer/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './AccountSelect';
 export * from './RepositorySearchBox';
+export * from './SearchRepository';
