+++ renderer/src/pages/SelectRepositoriesPage/SelectRepositoriesPage.tsx
@@ -1,16 +1,14 @@
 import React, { memo, useCallback, useState } from 'react';
 import { PlusCircleIcon, CheckCircleFillIcon } from '@primer/octicons-react';
-import { AccountSelect, RepositorySearchBox } from '../../components';
+import { SearchRepository } from '../../components';
 import { useSettings } from '../../hooks';
 import { BaseLayout } from '../../layouts/BaseLayout';
-import { searchRepositories } from '../../lib/searchRepositories';
 import {
   containerStyle,
   iconButtonStyle,
   iconStyle,
   repositoryListItemStyle,
   repositoryListStyle,
-  searchContainerStyle,
   titleStyle,
 } from './styles.css';
 
@@ -21,31 +19,19 @@ import {
  */
 
 export const SelectRepositoriesPage = () => {
-  const [account, setAccount] = useState<string | null>(null);
   const [repositories, setRepositories] = useState<string[]>([]);
   const { settings, addSubscribedRepository, removeSubscribedRepository } =
     useSettings();
 
-  const handleSearch = useCallback(
-    async (keyword: string) => {
-      const repositories = await searchRepositories(account ?? '', keyword);
-      setRepositories(repositories);
-    },
-    [account]
-  );
-
-  const handleAccountSelect = useCallback((account: string) => {
-    setAccount(account);
+  const handleSearch = useCallback(async (repositories: string[]) => {
+    setRepositories(repositories);
   }, []);
 
   return (
     <BaseLayout>
       <div className={containerStyle}>
         <h1 className={titleStyle}>リポジトリを選択</h1>
-        <div className={searchContainerStyle}>
-          <AccountSelect onSelect={handleAccountSelect} />
-          <RepositorySearchBox onSearch={handleSearch} />
-        </div>
+        <SearchRepository onSearch={handleSearch} />
         <RepositoryList
           repositories={repositories}
           subscribedRepositories={settings.subscribedRepositories}
